+++ internal/app/http/common.go
@@ -271,14 +271,17 @@ func convertParams(params any) *model.Filters {
 		&model.FilterOptions{NameFilterType: model.PrefixMatch})
 }
 
-func toInventoryResponse(ctx context.Context, res model.SearchResult) server.InventoryResponse {
+func toInventoryResponse(ctx context.Context, res model.SearchResult, page, pageSize int) server.InventoryResponse {
 	mapper := NewMapper(ctx)
 
 	meta := mapper.GetInventoryMeta(res)
 	inv := mapper.GetInventoryData(res.Entries)
 	resp := server.InventoryResponse{
-		Meta: &meta,
-		Data: inv,
+		Meta:       &meta,
+		Data:       inv,
+		TotalCount: &res.TotalCount,
+		Page:       &page,
+		PageSize:   &pageSize,
 	}
 	return resp
 }