+++ internal/app/http/service_test.go
@@ -168,7 +168,7 @@ func Test_ListInventory(t *testing.T) {
 		rMocks.MockReposAll(t, rMocks.CreateMockAllFunction(nil, r))
 		rMocks.MockReposGetDescriptions(t, []model.RepoDescription{{Name: "r1"}}, nil)
 		// when: list all
-		res, err := underTest.ListInventory(context.Background(), "", filters)
+		res, err := underTest.ListInventory(context.Background(), r.Spec().RepoName(), filters, 0, 0)
 		// then: there is no error
 		assert.NoError(t, err)
 		// and then: the search result is returned
@@ -184,7 +184,7 @@ func Test_ListInventory(t *testing.T) {
 		r2.On("Spec").Return(model.NewRepoSpec("r2")).Once()
 		rMocks.MockReposAll(t, rMocks.CreateMockAllFunction(nil, r, r2))
 		// when: list all
-		res, err := underTest.ListInventory(context.Background(), "", sp)
+		res, err := underTest.ListInventory(context.Background(), r.Spec().RepoName(), sp, 0, 0)
 		// then: there is an error of type repos.RepoAccessError
 		var aErr *repos.RepoAccessError
 		assert.ErrorAs(t, err, &aErr)